学校管理

+++ src/main/java/com/ruoyi/domain/SchoolLevel.java
@@ -17,7 +17,7 @@ public class SchoolLevel implements Serializable {
 
     private Integer schoolFeatureId;
 
-    private Long schoolId;
+    private Integer schoolId;
 
     private String schoolName;
